build(web): upgrade vite 7 → 8 with Rolldown-compatible fixes
Bumps vite to ^8.0.13 and @vitejs/plugin-react to ^6.0.2 (required for Vite 8).
Also bumps @tailwindcss/vite to ^4.3.0 and vitest catalog to 4.1.6 to satisfy
Vite 8 peer ranges without warnings.

Rolldown (Vite 8's bundler) enforces TanStack Start's import-protection plugin
more strictly than Rollup did, exposing two issues:

1. session.ts was statically importing @tanstack/react-start/server at module
   level, which the import-protection plugin blocks in the client bundle. Fixed
   by splitting into session.server.ts (server-only cookie logic) and a thin
   createServerFn wrapper in session.ts. API routes and tRPC now import from
   session.server directly; route beforeLoad/loader use the createServerFn path.

2. TanStack Start plugins pulled from vite.config.ts conflict with Storybook's
   multi-entry build under Rolldown. Fixed by filtering all tanstack* plugins
   out of the Storybook viteFinal config (stories only need React component
   context, not TanStack Start routing).
