+++ apps/web/package.json
@@ -68,14 +68,14 @@
     "@storybook/addon-docs": "catalog:storybook10",
     "@storybook/addon-vitest": "catalog:storybook10",
     "@storybook/react-vite": "catalog:storybook10",
-    "@tailwindcss/vite": "^4.1.18",
+    "@tailwindcss/vite": "^4.3.0",
     "@tanstack/react-query-devtools": "catalog:react-query5",
     "@tanstack/router-plugin": "catalog:tanstack-start",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "catalog:",
     "@types/react": "catalog:react",
     "@types/react-dom": "catalog:react",
-    "@vitejs/plugin-react": "^4.7.0",
+    "@vitejs/plugin-react": "^6.0.2",
     "@vitest/browser": "catalog:vitest",
     "@vitest/browser-playwright": "catalog:vitest",
     "@vitest/coverage-v8": "catalog:vitest",
@@ -95,7 +95,7 @@
     "testcontainers": "^11.11.0",
     "tw-animate-css": "^1.4.0",
     "typescript": "catalog:",
-    "vite": "^7.3.2",
+    "vite": "^8.0.13",
     "vite-tsconfig-paths": "^5.1.4",
     "vitest": "catalog:vitest",
     "vitest-mock-extended": "^3.1.0"
