+++ pnpm-workspace.yaml
@@ -64,10 +64,10 @@ catalogs:
     '@trpc/server': ^11.16.0
     '@trpc/tanstack-react-query': ^11.16.0
   vitest:
-    '@vitest/browser': 4.0.8
-    '@vitest/browser-playwright': 4.0.8
-    '@vitest/coverage-v8': 4.0.8
-    vitest: 4.0.8
+    '@vitest/browser': 4.1.6
+    '@vitest/browser-playwright': 4.1.6
+    '@vitest/coverage-v8': 4.1.6
+    vitest: 4.1.6
 
 linkWorkspacePackages: true
